Create a linear regression model, cross-validate it, and use it to predict the organic carbon content of soil samples based on their reflectance

Create a linear regression model, cross-validate it, and use it to predict the organic carbon content of soil samples based on their reflectance

cpred(data, reflectance, carbon, test_size, folds, prediction)

Source

The data was collected by Kangyu So, Dr. Alemu Gonsamo, and Dr. Cheryl A. Rogers (All affiliated with McMaster University) from the Hudson Bay Lowlands region on July 2023

Arguments

data

The dataset containing the average reflectance and organic carbon content of soil samples

reflectance

The average reflectance at a given wavelength range

carbon

The organic carbon content by percent dry weight

test_size

The test size for the model

folds

The number of folds for the cross-validation

prediction

The name for the predicted organic carbon content output

Value

A column containing the predicted organic carbon content using a cross-validated linear regression model

References

https://github.com/KangyuSo/my-GEO712-repository

Examples

data <- data.frame(ref_dry_avg)
pred_c <- cpred(data, "Reflectance", "Organic_Carbon", 0.7, 10, "Predicted_Organic_Carbon")
#> Loading required package: ggplot2
#> Loading required package: lattice