The function first extracts an algebraic equation based on the reflectance and predicted organic carbon content of soil samples. Then the function applies min-max normalization over the reflectance values of an Region of Interest (ROI) data table collected from image processing and analysis software ENVI. Then the function applies the extracted algebraic equation on the reflectance values of points in the ROI data file to predict organic carbon content. Then the function updates the ROI data file with the predictions and outputs it as a new csv file.

The function first extracts an algebraic equation based on the reflectance and predicted organic carbon content of soil samples. Then the function applies min-max normalization over the reflectance values of an Region of Interest (ROI) data table collected from image processing and analysis software ENVI. Then the function applies the extracted algebraic equation on the reflectance values of points in the ROI data file to predict organic carbon content. Then the function updates the ROI data file with the predictions and outputs it as a new csv file.

mappred(
predictions_dataframe, reflectance_column, prediction_column, 
ROI_dataset_path, output_dataset
)

Source

The data used for the predictions was collected by Kangyu So, Dr. Alemu Gonsamo, and Dr. Cheryl A. Rogers (All affiliated with McMaster University) from the Hudson Bay Lowlands region on July 2022. The ROI dataset was collected from images collected by the Moderate Resolution Imaging Spectroradiometer (MODIS) satellite operated by the National Aeronautics and Space Administration (NASA).

Arguments

predictions_dataframe

The dataset resulting from the output of the cpred function, contains the average reflectance and measured and predicted organic carbon content of some soil samples

reflectance_column

The average reflectance at a given wavelength range in the predictions_dataframe

prediction_column

The predicted organic carbon content in the predictions_dataframe

ROI_dataset_path

The path to the ROI data table that was exported from ENVI into a CSV file, contains geoinformation and reflectance values of the ROI divided into individual points

output_dataset

The updated ROI data table containing the new predicted organic carbon content of the points

Value

A CSV file containing the ROI data table as well as the predicted organic carbon content of the points using an algebraic equation extracted from the predictions of a cross-validated linear regression model

References

https://github.com/KangyuSo/carbcontpred

Examples

mappred_c <- mappred(
pred_c, "Reflectance", "Predicted_Organic_Carbon", 
system.file("extdata", "Reflectance_ROI.csv", package = "carbcontpred"), "ROI_Predicted_Carbon"
)
#> Rows: 12209 Columns: 7
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (1): ; File X
#> dbl (6): File Y, Map X, Map Y, Lat, Lon, B1
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Error in mappred(pred_c, "Reflectance", "Predicted_Organic_Carbon", system.file("extdata",     "Reflectance_ROI.csv", package = "carbcontpred"), "ROI_Predicted_Carbon"): object 'ROI_folder' not found